✨ 関数分けした
